% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_function}
\alias{call_function}
\title{Call a function by character strings using the namespace and custom 
parameters.}
\usage{
call_function(func_list)
}
\arguments{
\item{func_list}{A list with fields func, namespace, and args}
}
\description{
Call a function by character strings using the namespace and custom 
parameters.
}
