% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRBase.R
\docType{data}
\name{CBRBase}
\alias{CBRBase}
\title{Root class for common functionality of this package}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CBRBase
}
\description{
Root class for common functionality of this package
}
\keyword{c++}
\keyword{data}
\keyword{data-preparation}
\keyword{for}
\keyword{functions}
\keyword{integer}
\keyword{necessary}
\keyword{representation;}
\keyword{to}
\keyword{transforms}
