% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxBetaModel.R
\docType{class}
\name{CoxBetaModel}
\alias{CoxBetaModel}
\title{Cox-Beta Model}
\format{An \code{\link{R6Class}} generator object}
\usage{
CoxBetaModel
}
\description{
Regression beta coefficients are use for building a weighted distance measure between
the learning and verum data set. The learning data set is used for learning the Cox
model and use the obtained weights for calculating a (n x m)-distance
matrix, where n is the number of cases in the learning data set and m is the
number of cases of the query data. This distance matrix can then be used for
cluster analysis or for getting for each case in the query data k (=1,...,l)
smilar cases from the learning data. The rms-package is used for model fitting,
variable selection, and checking the assumptions.
If query data is ommitted, a n x n- distance matrix is returned.
}
\section{Usage}{

For usage details see \bold{Methods, Arguments, and Examples} sections.
}

\section{Methods}{

\describe{
  \item{\code{new(formula, ...)}}{This method is used to create an
  object of this class \code{RFModel}. Formula for analysis has to be 
  provided. Further parameters for the ranger call can be provided here.}
  \item{\code{fit(dtData)}}{Fits the RandomForest model.}
  \item{...}{See \link{CBRBase} class.}
  }
}

\keyword{datasets}
