% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{depthMatrix}
\alias{depthMatrix}
\title{Get depth distance matrix}
\usage{
depthMatrix(x, y = NULL, rfObject)
}
\arguments{
\item{x}{a new dataset}

\item{y}{a new dataset}

\item{rfObject}{\code{ranger} object}
}
\description{
Get depth distance matrix
}
\examples{
require(ranger)
rf <- ranger(Species ~ ., data = iris, num.trees = 5, write.forest = TRUE)
d <- depthMatrix(x=iris[, -5], rfObject=rf)

}
