% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{edges_between_terminal_nodes}
\alias{edges_between_terminal_nodes}
\title{Number of Edges between Terminal Nodes}
\usage{
edges_between_terminal_nodes(rfObject)
}
\arguments{
\item{rfObject}{\code{ranger} object}
}
\value{
a \code{matrix} object with pairwise terminal node edge length
}
\description{
first two columns are terminal node IDs; If an ID pair do not 
appear in a tree -1 is inserted
}
\examples{
\donttest{
require(ranger)
rf.fit <- ranger(Species ~ ., data = iris, num.trees = 5, write.forest = TRUE)
edges_between_terminal_nodes(rf.fit)
}

}
