% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionModel.R
\name{LogisticModel}
\alias{LogisticModel}
\title{Logistic Regression Model for Case-Based-Reasoning}
\description{
Logistic Regression Model for Case-Based-Reasoning

Logistic Regression Model for Case-Based-Reasoning
}
\section{Super classes}{
\code{\link[CaseBasedReasoning:CBRBase]{CaseBasedReasoning::CBRBase}} -> \code{\link[CaseBasedReasoning:RegressionModel]{CaseBasedReasoning::RegressionModel}} -> \code{LogisticModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{the statistical model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LogisticModel-clone}{\code{LogisticModel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="calc_distance_matrix"><a href='../../CaseBasedReasoning/html/CBRBase.html#method-CBRBase-calc_distance_matrix'><code>CaseBasedReasoning::CBRBase$calc_distance_matrix()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="get_similar_cases"><a href='../../CaseBasedReasoning/html/CBRBase.html#method-CBRBase-get_similar_cases'><code>CaseBasedReasoning::CBRBase$get_similar_cases()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="initialize"><a href='../../CaseBasedReasoning/html/CBRBase.html#method-CBRBase-initialize'><code>CaseBasedReasoning::CBRBase$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="RegressionModel" data-id="fit"><a href='../../CaseBasedReasoning/html/RegressionModel.html#method-RegressionModel-fit'><code>CaseBasedReasoning::RegressionModel$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="RegressionModel" data-id="print"><a href='../../CaseBasedReasoning/html/RegressionModel.html#method-RegressionModel-print'><code>CaseBasedReasoning::RegressionModel$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="RegressionModel" data-id="variable_selection"><a href='../../CaseBasedReasoning/html/RegressionModel.html#method-RegressionModel-variable_selection'><code>CaseBasedReasoning::RegressionModel$variable_selection()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
