% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionModel.R
\name{RegressionModel}
\alias{RegressionModel}
\title{Root class for Regression Models, e.g., CPH, logistic, and linear regression}
\description{
Root class for Regression Models, e.g., CPH, logistic, and linear regression

Root class for Regression Models, e.g., CPH, logistic, and linear regression
}
\keyword{data-preparation}
\section{Super class}{
\code{\link[CaseBasedReasoning:CBRBase]{CaseBasedReasoning::CBRBase}} -> \code{RegressionModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_params}}{rms arguments}

\item{\code{weights}}{Weights for distance calculation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RegressionModel-print}{\code{RegressionModel$print()}}
\item \href{#method-RegressionModel-variable_selection}{\code{RegressionModel$variable_selection()}}
\item \href{#method-RegressionModel-fit}{\code{RegressionModel$fit()}}
\item \href{#method-RegressionModel-clone}{\code{RegressionModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="calc_distance_matrix"><a href='../../CaseBasedReasoning/html/CBRBase.html#method-CBRBase-calc_distance_matrix'><code>CaseBasedReasoning::CBRBase$calc_distance_matrix()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="get_similar_cases"><a href='../../CaseBasedReasoning/html/CBRBase.html#method-CBRBase-get_similar_cases'><code>CaseBasedReasoning::CBRBase$get_similar_cases()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="initialize"><a href='../../CaseBasedReasoning/html/CBRBase.html#method-CBRBase-initialize'><code>CaseBasedReasoning::CBRBase$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionModel-print}{}}}
\subsection{Method \code{print()}}{
Prints information of the initialized object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionModel$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionModel-variable_selection"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionModel-variable_selection}{}}}
\subsection{Method \code{variable_selection()}}{
Fast backward variable selection with penalization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionModel$variable_selection(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Training data of class data.frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionModel-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the RandomForest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionModel$fit()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Training data of class data.frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
