% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\docType{class}
\name{Validate}
\alias{Validate}
\title{R6 Validation Class for case based reasoning}
\format{An \code{\link{R6Class}} generator object}
\usage{
Validate
}
\description{
R6 Validation Class for case based reasoning
}
\keyword{validation}
