% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{weightedDistance}
\alias{weightedDistance}
\title{Weighted Distance calculation}
\usage{
weightedDistance(x, y = NULL, weights = NULL)
}
\arguments{
\item{x}{a new dataset}

\item{y}{a second new dataset}

\item{weights}{a vector of weights}
}
\value{
a \code{dist} or \code{matrix} object
}
\description{
Weighted Distance calculation
}
\examples{
\donttest{
require(ranger)
rf <- ranger(Species ~ ., data = iris, num.trees = 5, write.forest = TRUE)
terminalNodes(iris[, -5], rf)
}

}
